Split objects
