+++ LBH.AdultSocialCare.Api/V1/Domain/HomeCare/HomeCarePackageSlotListDomain.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
 
 namespace LBH.AdultSocialCare.Api.V1.Domain.HomeCare
 {
@@ -17,51 +16,6 @@ namespace LBH.AdultSocialCare.Api.V1.Domain.HomeCare
         /// Gets or sets the Home Care Package Slots
         /// </summary>
         public List<HomeCarePackageSlotDomain> HomeCarePackageSlots { get; set; }
-
-    }
-
-    public class HomeCarePackageSlotDomain
-    {
-
-        /// <summary>
-        /// Gets or sets the primary in minutes.
-        /// </summary>
-        public int PrimaryInMinutes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the secondary in minutes.
-        /// </summary>
-        public int SecondaryInMinutes { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Time Slot Shift Id
-        /// </summary>
-        public int TimeSlotShiftId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Time Slot Shift
-        /// </summary>
-        public TimeSlotShifts TimeSlotShift { get; set; }
-
-        /// <summary>
-        /// Gets or sets the day identifier.
-        /// </summary>
-        public int DayId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Need To Address
-        /// </summary>
-        public string NeedToAddress { get; set; }
-
-        /// <summary>
-        /// Gets or sets the What Should Be Done
-        /// </summary>
-        public string WhatShouldBeDone { get; set; }
-
-        /// <summary>
-        /// Gets or sets the Service Id
-        /// </summary>
-        public Guid ServiceId { get; set; }
 
     }
